// generated/src/app/api-models/case-reason.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CaseReason {
  seqCaseId: number;
  seqCaseReasonId: number;
  seqCallerId: number;
  sequenceNumber: number;
  transDate: Date;
  reasonCsCode: string;
  reasonCat1CsCode: string;
  reasonCat2CsCode: string;
  contactToneCsCode: string;
  userDefined1: string;
  userDefined2: string;
  userDefinedDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}